feat: Add create BLoC feature
- Added create BLoC command.
- Changed application structure.
- Refactored code to separate feature and BLoC commands into distinct files.
- Updated package.json to reflect new command and context menu items.
